requirements.txt
fastapi
uvicorn
httpx
python-dotenv
tenacity>=8.0.0 